requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
psycopg2-binary
